requirements.txt
faiss-cpu
sentence-transformers
requests
pypdf
numpy
